Force refresh should scan for deleted extents too
It's supposed to be the "just recreate everything" option,
for when someone has messed with their postgres database manually.

A few people on Slack (including myself) found it confusing
that you need to include a second option in addition to `--force`
to scan for deleted datasets. We should just include that with
`--force`.