Improve CRS inference
The test is broken on newer docker container -- built-in WKT strings
are slightly different.

Strangely, it was sorting ascending and picking the first. ... so the least
good match? But this worked because the test data used a completely different capitalisation to the real CRS, so they didn't match at all.

Changed them to:

1. Compare without case sensitivity.
2. Compare total matches, rather than ratio. WKTs are very diff.
   lengths, and we want to match their strings of names/numbers,
   not penalize for missing sections (which is common).
3. Pick the best match, not the least