docs: fix flask-caching URL
The URL gives me a "This site can't be reached"
message, stating "might be temporarily down or
it may have moved permanently to a new web
address" with a ERR_INVALID_RESPONSE below.

Trying with curl shows a 404. Rather than
trying to figure out what once was at the page
and how to access that content, update
the URL to point to the documentation
at readthedocs.io that just works.