CI: pass token to CodeCov action
This repository lags behind with
the CodeCov baseline for some reason,
and the other ODC repositories are
usually close. Paul has a memory
of having had the same problem
in some other repository and suggested
it might be the missing token.

Align the parameters with datacube-core
where CodeCov works as we want, so pass
in the token to the workflow and disable
the search for additional coverage files.
The disabled search shaves a couple of
seconds off a couple of seconds of runtime
of the action itself and has nothing
to do with the lagging baseline.