stores: use CreateSchema
There is already a DropSchema in
stores, so put the CreateSchema
in there as well.

The previous code has some comment
about being able to run things
without create permissions, but
the only caller is SummaryStore.init()
whose documentation states that
it requires create permissions,
so use the if_not_exists parameter
to CreateSchema instead of some
function that probes the database.