generate: always use Pool
The imap implementation will be
as lazy as one could expect in
Python 3.15, but until then
it is more eager than one would
expect.

Switch from imap_unordered to
using apply_async instead, but
print the results in a deterministic
order and remove the special
case where a different code path
is taken when running on a single
core.