Fix version number in Docker builds
We were getting an unclean git repo inside of the Docker build
which is where setuptools_scm determines the version number.
This happend because our .dockerignore was excluding some directories that
exist in the git repo but aren't required in the docker image.

That's reasonable, but it makes it impossible to determine the version
number correctly inside of the docker build.

This change add a docker ARG, and determines the version number in the GitHub
Actions workflow before building the image.

Closes #897